REFACTOR : Issue #117 Phase 3 Commit 1 - move v2 stokes/mantle/slab/forcing-{slab_pull,mantle_force}/rheology/basal_drag/presets to _attic, gated by v2_legacy feature
Per docs/migrations/v2_to_c1_attic.md audit + user-validated HC1+HC2+HC4
+ SC2 decisions, this commit physically moves the Stokes-coupled v2
subtree under tectonics_v2/_attic/ and gates it behind Cargo feature
v2_legacy. The default workspace build excludes the retired surface;
cargo build --workspace --features v2_legacy rebuilds it for
regression reproducibility (R5b mf sweeps, R6.3 evo sweep, R7.A.2.4
init mode sweep).

Moves (36 files):

  tectonics_v2/stokes/                          (22 files: AMG subtree
                                                 + solver, picard,
                                                 nonlinear_solver,
                                                 operator, snapshot,
                                                 sparse_assembly,
                                                 nullspace, precond,
                                                 parallel_reduce,
                                                 continuation, mod)
  tectonics_v2/mantle/                          (3 files: mod,
                                                 pattern, stream_function)
  tectonics_v2/slab/                            (4 files: mod, state,
                                                 accumulation,
                                                 convergence_direction)
  tectonics_v2/forcing/{slab_pull,mantle_force}.rs (2 files)
  tectonics_v2/rheology.rs
  tectonics_v2/basal_drag.rs
  tectonics_v2/presets.rs

  All moved into tectonics_v2/_attic/<subtree>/ keeping the source
  structure. New _attic/mod.rs and _attic/forcing/mod.rs declare the
  retired submodules; both gate with #![cfg(feature = "v2_legacy")].

Path fixes in moved files (mechanical, ~25 lines across ~22 files):

  super::super::X (in subtree files, X stays in tectonics_v2)
    -> crate::tectonics_v2::X
  super::X (in top-level moved files, X stays in tectonics_v2)
    -> crate::tectonics_v2::X

  Paths where target also moved to _attic stay valid (super::super
  walks the module tree, both source and target relocated as a unit).
  Specifically preserved: super::super::{stokes,mantle,slab,rheology,
  basal_drag,presets} and super::super::forcing::{slab_pull,
  mantle_force} which all resolve correctly post-move.

Parent mod.rs updates:

  tectonics_v2/mod.rs                  removes pub mod {stokes,mantle,
                                       slab,rheology,basal_drag,
                                       presets}; declares cfg-gated
                                       pub mod _attic + re-exports
                                       (pub use _attic::{...} +
                                       pub use _attic::rheology::
                                       soft_min_harmonic)
  tectonics_v2/forcing/mod.rs          removes pub mod {slab_pull,
                                       mantle_force} + re-exports;
                                       adds cfg-gated re-export
                                       from super::_attic::forcing
  tectonics_v2/diagnostics/mod.rs      gates pub mod {ar_sweep,
                                       bi_sweep,br_sweep,harness,
                                       k_sub_sweep,mms_bench,
                                       num_plates_sweep,report} +
                                       re-exports (harness items
                                       transitively gated; report
                                       gated because it aggregates
                                       sweep renderers)
  tectonics_v2/workflow/mod.rs         gates pub mod {phase_a,
                                       phase_b} + their re-exports +
                                       CycleOutput + PhaseAOutput
                                       (carry BaselineResult)

Cargo.toml updates:

  crates/ymir-core/Cargo.toml          new [features] v2_legacy = [];
                                       bench amg_benchmark +
                                       bins {step_baseline,step5-8_
                                       baseline,gen_reference_
                                       solutions,gen_bench_data}
                                       gated via required-features
  crates/ymir-viz/Cargo.toml           v2_legacy = ["ymir-core/v2_
                                       legacy"]; bin ymir-viz gated
                                       via required-features
  crates/ymir-viz/src/lib.rs           full-file gate (the entire
                                       v2 bridge + UI panels + v2_viz
                                       chain is Stokes-coupled; C1
                                       will reintroduce a paradigm-
                                       agnostic facade in C1 Phase 4)

Build verification:

  cargo check --workspace               OK (default, no feature)
  cargo check --workspace --features v2_legacy  OK
  All 36 retired files detected as renames (git status RM)
  Path fixes: 0 super::* remaining that resolve outside _attic without
  a crate:: prefix

NOT yet in this commit (Commit 2 scope):

  - Test gating / relocation to tests/_attic/ (SC2 confirmed)
  - cargo test --workspace default still fails on retired-importing tests
  - Bridge/UI granular gating (currently the entire ymir-viz lib gates;
    SC4 spec roundtrip without feature is not yet preserved)

Commit 2 will gate or relocate the ~32 test files identified in the
audit (D5 inventory), restoring cargo test --workspace default green.
