+++ crates/ymir-core/src/tectonics_v2/_attic/stokes/amg/coarse_solve.rs
