+++ crates/ymir-core/src/tectonics_v2/_attic/stokes/amg/restriction.rs
