+++ crates/ymir-core/src/tectonics_v2/_attic/stokes/nullspace.rs
