+++ crates/ymir-core/src/tectonics_v2/_attic/stokes/snapshot.rs
